requirements.txt
##dependencias
#app
flask
flask-cors
flask-sqlalchemy
flask-marshmallow
flask_jwt_extended
marshmallow-sqlalchemy
Werkzeug
psycopg2-binary
pymongo[srv,aws]
bson
python-dotenv
boto3

#tasks
celery
redis

#deploy
gunicorn
wheel
